+++ Warehouse.php
@@ -328,6 +328,20 @@ if ( ! empty( $_REQUEST['modfunc'] )
 		|| $_SERVER['HTTP_X_CSRF_TOKEN'] !== $_SESSION['token'] )
 	{
 		// Security fix #371 check CSRF token for modfunc requests
+		if ( empty( $_SESSION['STAFF_ID'] ) && empty( $_SESSION['STUDENT_ID'] ) )
+		{
+			// No user in session yet, 403 Forbidden error (99% chances its a bot)
+			http_response_code( 403 );
+
+			$error[] = _( 'You\'re not allowed to use this program!' );
+
+			ErrorSendEmail( $error, 'HACKING ATTEMPT' );
+
+			echo ErrorMessage( $error );
+
+			exit;
+		}
+
 		require_once 'ProgramFunctions/HackingLog.fnc.php';
 
 		HackingLog();